+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -290,7 +290,7 @@ impl WalletMessageHandler {
                             let node_info = Client::get_node_info(&url, auth).await?;
                             Ok(Response::NodeInfo(NodeInfoWrapper { node_info, url }))
                         }
-                        None => self.wallet.get_node_info().await.map(Response::NodeInfo),
+                        None => Ok(self.wallet.client().get_info().await.map(Response::NodeInfo)?),
                     }
                 })
                 .await
@@ -356,10 +356,12 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let bech32_hrp = match bech32_hrp {
                         Some(bech32_hrp) => bech32_hrp,
-                        None => match self.wallet.get_node_info().await {
-                            Ok(node_info_wrapper) => node_info_wrapper.node_info.protocol.bech32_hrp,
-                            Err(_) => SHIMMER_TESTNET_BECH32_HRP,
-                        },
+                        None => self
+                            .wallet
+                            .client()
+                            .get_bech32_hrp()
+                            .await
+                            .unwrap_or(SHIMMER_TESTNET_BECH32_HRP),
                     };
 
                     Ok(Response::Bech32Address(utils::hex_to_bech32(&hex, bech32_hrp)?))