+++ sdk/CHANGELOG.md
@@ -105,6 +105,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Bech32Addresses` and `RawAddresses`;
 - `Client::get_addresses`;
 - `Topic` `TryFrom<String>` impl;
+- `Client::generate_ed25519_addresses`
+- `Wallet::get_node_info`
 
 ### Fixed